requirements.txt
absl-py==0.10.0
astor==0.8.1
astunparse==1.6.3
bleach==1.5.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
flatbuffers==1.12
gast==0.3.3
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
html5lib==0.9999999
idna==2.10
importlib-metadata==2.0.0
Keras==2.4.3
Keras-Applications==1.0.8
keras-contrib==2.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
nibabel==3.1.1
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tb-nightly==2.4.0a20201021
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
tf-estimator-nightly==2.4.0.dev2020102101
tf-nightly-gpu==2.4.0.dev20201022
typing-extensions==3.7.4.3
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.2.0
